requirements.txt
# Tests
pylint
pytest
pytest-cov

# Utilities
tqdm
boto3==1.13.12
pyyaml
dicttoxml
xmltodict
PyGithub

# plotting
matplotlib

# Data
numpy
pandas
scipy

# Computer Vision
Pillow
opencv-python==4.2.0.34

# Internal APIs
fuegosecrets@git+ssh://git@github.pie.apple.com/RetailAppsATX/Fuego-Secrets.git@v2.5.0

# For Tag s3 upload
python-dotenv
turi-tag-keycloak-connector==1.9.0

# Turi Trove for public datasets
turicreate==6
